requirements.txt
aiohttp==3.7.4.post0
astroid==2.9.3
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord.py==1.7.3
dnspython==2.2.0
flake8==4.0.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
motor==2.5.1
multidict==6.0.2
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
pymongo==3.12.3
python-dotenv==0.19.2
toml==0.10.2
typing_extensions==4.1.1
wrapt==1.13.3
yarl==1.7.2
